--- pages/404.tsx
@@ -1,3 +1,0 @@
-export default function My404(){
-    return <h1>404 ㅠㅠ</h1>
-}
