--- pages/_document.tsx
@@ -1,10 +1,0 @@
-import NextDocument from 'next/document'
-import {createGetInitialProps} from '@mantine/next'
-
-const getInitialProps = createGetInitialProps()
-
-export default class _Document extends NextDocument {
-    static getInitialProps = getInitialProps
-}
-
-
