+++ src/app/posts/[...id].tsx
