--- styles/utils.module.css
@@ -1,64 +1,0 @@
-.heading2Xl {
-    font-size: 2.5rem;
-    line-height: 1.2;
-    font-weight: 800;
-    letter-spacing: -0.05rem;
-    margin: 1rem 0 0 0;
-}
-
-.headingXl {
-    font-size: 2rem;
-    line-height: 1.3;
-    font-weight: 800;
-    letter-spacing: -0.05rem;
-    margin: 1rem 0;
-}
-
-.headingLg {
-    font-size: 1.5rem;
-    line-height: 1.4;
-    margin: 1rem 0;
-}
-
-.headingMd {
-    font-size: 1.2rem;
-    line-height: 1.5;
-    margin: 0.5rem 0;
-}
-
-.borderCircle {
-    border-radius: 9999px;
-}
-
-.colorInherit {
-    color: inherit;
-}
-
-.padding1px {
-    padding-top: 1px;
-}
-
-.list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.listItem {
-    margin: 0 0 1.25rem;
-}
-
-.lightText {
-    color: #666;
-}
-
-.pointESC {
-    -webkit-text-fill-color: transparent;
-    color: coral;
-    background-image: url('/images/rainbow2.jpeg');
-    -webkit-background-clip: text;
-    background-clip: text;
-    background-size: 1em;
-    font-size: 1em;
-    font-weight: bold;
-}
